- gui WIP

+++ src/rez/cli/_util.py
@@ -95,6 +95,10 @@ _handled_int = False
 _handled_term = False
 
 
+def _env_var_true(name):
+    return (os.getenv(name, "").lower() in ("1", "true", "on", "yes"))
+
+
 def sigbase_handler(signum, frame):
     # show cursor - progress lib may have hidden it
     SHOW_CURSOR = '\x1b[?25h'
@@ -102,7 +106,9 @@ def sigbase_handler(signum, frame):
     sys.stdout.flush()
 
     # kill all child procs
-    os.killpg(os.getpgid(0), signum)
+    # FIXME this kills parent procs as well
+    if not _env_var_true("_REZ_NO_KILLPG"):
+        os.killpg(os.getpgid(0), signum)
     sys.exit(1)
 
 
@@ -111,7 +117,8 @@ def sigint_handler(signum, frame):
     global _handled_int
     if not _handled_int:
         _handled_int = True
-        print >> sys.stderr, "Interrupted by user"
+        if not _env_var_true("_REZ_QUIET_ON_SIG"):
+            print >> sys.stderr, "Interrupted by user"
         sigbase_handler(signum, frame)
 
 
@@ -120,7 +127,8 @@ def sigterm_handler(signum, frame):
     global _handled_term
     if not _handled_term:
         _handled_term = True
-        print >> sys.stderr, "Terminated by user"
+        if not _env_var_true("_REZ_QUIET_ON_SIG"):
+            print >> sys.stderr, "Terminated by user"
         sigbase_handler(signum, frame)
